Mario Test Update 3
I have all three pickups that the tutorial covers implemented and tested.

There is a small issue with the shrinking of mario in which he can remain against the enemy  until he breaks contact and re-touches the enemy, if I remember correct the enemy should pass through Mario for a short period of time after he returns to his small state.